Add shared nav shell via route group
