+++ src/app/(app)/reports/page.tsx
@@ -1,0 +1,12 @@
+import PageHeader from "@/app/components/PageHeader";
+
+export default function ReportsPage() {
+  return (
+    <>
+      <PageHeader title="Reports" />
+      <div className="p-6 max-w-7xl mx-auto w-full">
+        <p className="text-gray-500">Reports are coming soon.</p>
+      </div>
+    </>
+  );
+}
