--- src/app/dashboard/layout.tsx
@@ -1,7 +1,0 @@
-export default async function DashboardLayout({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
-  return <div className="min-h-screen bg-gray-50">{children}</div>;
-}
